Implement some of team
This required adding the ability to delete. I moved a bunch of
common model object code to a base object which can be utilized
with composition. I also created a model for roster_spot.

Right now, team model is building its own query for roster because
it requires joins, which is not supported by the query builder.
This will become a pain and a violation of DRY moving forward,
so my next task ought to be building that capability.