Refactor options builder such that conditions and fields are separate
It made no sense to include conditions and fields together in the same
object. An object was being passed around that had pieces that
weren't always relevant to its use.

I also modified the query builders so that they never take
in a string. It didn't make sense to generate part of a query
and then pass that in to the query builder to generate the
rest of it. My goal is to move all references to the subquery
builders into the query builders. In order to make that happen,
I need to find a way to do joins within the select query builder.